Add check for fault equivalence (#2)
Adds a function that compares two noise models for fault equivalence.
This is an adaptation of the algorithm developed in
https://github.com/maximilianruesch/fault-gadgets.

As such, only equally and normally weighted noise models are supported.
Further, until we can correctly construct bijections between inputs and
outputs of diagrams, the boundary node IDs must be ordered the same in
the graphs, i.e. they must be inserted in the same order.
Finally, this version of the algorithm does not yet support direct
connections of boundary nodes, since boundary nodes must have a 1-1
allocation of boundary edges. This can be solved by binding Pauli strings to ZXIO instead of boundary edges.